[CW] [73045828] Rename Sinatra/Slim Helper
