+++ lib/pinball/rails/slim/helper.rb
@@ -1,0 +1,13 @@
+module Pinball
+  module Rails
+    module Slim
+      module Helper
+        def feature(name)
+          if Pinball::Registry.available?(name)
+            # Placeholder for future integration
+          end
+        end
+      end
+    end
+  end
+end
